Move ResetConfiguration from Help to Tools menu.
Requested by Peter, and I agreed.
